requirements.txt
openai>=1.0
python-dotenv>=1.0
streamlit>=1.28
pytest>=7.0
